Update Pulumi StackCRD and add execution components for Kubernetes
This commit updates the visibility of members in Pulumi StackCRD structure to public for external access. It also adds various components related to Pulumi Execution including StackSourceRepository, responsible for retrieval of StackSource, StackSourceRepository, and ClusterStackSource based on their names and namespace. Additional CRDs for ClusterStackSource and StackSource are now included in PulumiStackCrdInstaller. Extensions were made to KubernetesService with added methods for accessing resources based on their names and namespace. The image tag for the Kubernetes job under PulumiStackService was also updated.