Update pulumi command path and improve Docker configuration
Change the path for the 'pulumi' command in the Pulumi CLI lib from '/root/.pulumi/bin/pulumi' to 'pulumi' to enable better path lookup. Also, added 'npm' as an installed package in the Dockerfile for the pulumi-operator-kubernetes-job improving container build. The image for pulumi-operator-kubernetes-job in the Kubernetes Service Stack was updated to 1.0.10.

This refactoring improves the robustness of the code and ensures that the container has all necessities for execution.