Update Dockerfile for NVM installation and node version
The Dockerfile for the pulumi-operator-kubernetes-job has been updated to correct the installation of Node Version Manager (NVM) and also to switch to an older version of Node.js (v14.18.1). This is because the application relies on features that are incompatible with the latest node version, so we've decided to default to version 14 accordingly.