Add fetch_git_notes and adjust image version
Fetch function added to get git notes and update callbacks for the repository. This improves the GitController by allowing to authenticate and fetch git note references. The image version of the main_container in `pulumi-operator-kubernetes/src/stack/service.rs` has been updated from 1.0.20 to 1.0.21 for the pulumi operator kubernetes job.