Add main pod overrides and improve cronjob naming
Modified the service code to include main pod overrides directly from the stack's spec. This new feature allows users to override extra annotations in the main pod.
Additionally, changed CronJob name formatting, now the CronJobs will be prefixed with 'pulumi-', improving clarity in the naming of these jobs.
Also expanded the list of allowed verbs in the rules, to allow the CronJob more freedom, matching the level of freedom given to the initial containers.