Add support for organization in pulumi stack init
Added a feature to the Pulumi stack initialization to include organization details. This is implemented by introducing 'organization' as an option in both the 'StackInitOptions' struct and the Pulumi stack specification. Accordingly, the 'stack_init' function syncs the stack name with the present organization when initiating a new stack. Also, updated Pulumi Kubernetes job image to version 1.0.26. This feature enhances stack management in organizations by enabling namespace differentiation.